renamed inputs to item_types
