+++ pype/tools/settings/settings/widgets/__init__.py
@@ -1,10 +1,9 @@
 from .window import MainWidget
-# TODO properly register inputs to TypeToKlass class
-from . import inputs
-from . import anatomy_inputs
+from . import item_types
+from . import anatomy_type
 
 __all__ = [
     "MainWidget",
-    "inputs",
-    "anatomy_inputs"
+    "item_types",
+    "anatomy_type"
 ]
